Improvements: Cut decode speed by half, readme updates, unused things (#8)
Cuts the decode speed in half by moving all of the necessary accesses to the shuffle and length table into assembly code. This likely resulted in a 3 ns drop due to inlining of the fast func in the benchmark code.

Includes some various code tidies, e.g. removing unused things as well as some readme updates.